refactor: controllerTest -> E2ETest 명칭 변경
