+++ src/test/java/roomescape/e2e/ReservationApiE2ETest.java
@@ -19,7 +19,7 @@ import static org.hamcrest.CoreMatchers.is;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class ReservationApiTest {
+public class ReservationApiE2ETest {
 
     @Autowired
     private JdbcTemplate jdbcTemplate;
@@ -123,5 +123,19 @@ public class ReservationApiTest {
                 .then().log().all()
                 .statusCode(200)
                 .body("size()", is(2));
+    }
+
+    @DisplayName("존재하지 않는 예약을 삭제하려는 경우 404 Not Found를 던진다")
+    @Test
+    void reservationRemove() {
+        //given
+        long notExistId = 100L;
+
+        //when & then
+        RestAssured.given().log().all()
+                .contentType(ContentType.JSON)
+                .when().delete("/reservations/" + notExistId)
+                .then().log().all()
+                .statusCode(404);
     }
 }
